Allow datastore clusters to be nested in folders
Slashes are not valid in folder or datastore cluster names, they must be escaped in vCenter, so we use slash
as our expected folder delimiter.

For the "pattern" matching case, after splitting on slashes, we have to convert each segment into a regular expression
in case the user is relying on that for wildcard matching. As regexes don't use forward slashes, and the entity names they are
searching for also can't have slashes, this should be safe.

[#185046564]