NSXT Policy Client reties when searching for virtual machines
There seems to be a delay between when the VM has been created, and when
NSXT sees the VM. This can lead to failures with the error:

Failed to find vm in realized state with cid

When in fact the VM does exist. This occurs when the CPI tries to add
the VM to an NSXT Group, which is where VMs are placed when they are
load balancer endpoints.

We now retry the find.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>