Generate release tag when pushing release commit
Previously, the GitHub release resource was creating the tag when a new
release was cut. This could lead to a situation where Concourse would
pick up the new commit before the tag was created, and pass that
tag-less version of the repo to the `automatically-release-new-patch`
job, which would then fail because the tag didn't exist.

[#185631201] Ensure components using Golang are auto-released (as well as auto-bumped)