Do not pass block to Tempfile.create
When a block is passed to Tempfile.create, it returns whatever the last
line of the block was (in this case, the number of bytes written), not
the tempfile object. Additionally, it immediately deletes the file after
the block is executed. We are already doing cleanup of the files at the
end of executing the CPI.

Signed-off-by: Joseph Palermo <jpalermo@vmware.com>