Various fixes to the Docker image build
This commit:
* Stops building both amd64 and arm64 images, we now only build amd64
  images.
* Stops installing ginkgo into the Docker image

Why?

1) We switched to the 'bosh/golang-release' Docker image as our base, which doesn't have an ARM image, so attempts to build a combined image causes the following error:
```
 > [linux/arm64 3/8] RUN apt-get update && apt-get install -y --no-install-recommends   ca-certificates   curl   git   git-lfs   jq   make   neovim   netcat-openbsd   openssh-client   openvpn   rsync   sshpass   sshuttle   tar   wget   zsh     &&   rm -rf /var/lib/apt/lists/*:
```

2a) Switching base images has moved where 'go' is installed, so the
build process was trying to execute `/usr/local/go/bin/go install
github.com/onsi/ginkgo/v2/ginkgo@latest;`, but `go` wasn't at that path,
and it failed.

2b) Apparently, installing your build tools into system-level
directories is a super bad idea in this new world of Go Modules. The
idea is that you 'go run' them, which causes the specific version in
your 'go.mod' file to be downloaded and compiled, avoiding version
mismatches.

Signed-off-by: Kenneth Lakin <klakin@vmware.com>