Explicitly call out memory units (MB, GB)
Our `fetch_resource_pool_utilization` method has been broken:
`quick_stats.host_memory_usage` was treated as in kiB when it was in
bytes.

This didn't have much fallout; the miscalculation would've only
affected deployments to memory-starved resource pools, and the only
manifestation would've been to error later in the deployment phase
rather than earlier.

To avoid this type of error, we generously update the variables and
methods that return memory-related information to include the units
(typically MiB).