Enable expanding root disk `root_disk_size_gb`
Applications such as Nvidia's libraries, required for many AI / LLM
applications, aren't easily installed on anywhere but the root disk;
however, the amount of free space in the root disk of a Jammy stemcell
is too small to accommodate these libraries.

To address this, we introduce a new property, `root_disk_size_gb`.
Setting this property will expand the root disk during VM creation (in
the `create_vm` CPI method, before the BOSH packages are installed).

Typical use (in a Cloud Config):

```yaml
vm_extensions:
- name: 20G_root
  cloud_properties:
    root_disk_size_gb: 20
```

Setting this property disables the VM's "linked clone" feature [0],
which is incompatible with extending the size of the root disk. The VM
will therefore require more disk space than expected, typically the size
of the root disk (in the above example, 20 GiB)

Adding this feature brings the vSphere CPI to parity with the AWS,
Azure, and GCP CPIs.

Root disk size is measured in GiB (1,073,741,824 bytes), not in GB
(1,000,000,000 bytes).

Drive-by:

Removed unnecessary initialization, `config_spec.device_change = []`:
`ConfigSpec` class initializes `device_change` to an empty array; we
don't need to do it.

[0] A linked clone is a copy of a virtual machine that shares virtual
disks with the parent virtual machine. Only changes from the parent disk
are recorded, saving on disk space.

Signed-off-by: Brian Cunnie <brian.cunnie@broadcom.com>