Restore IP conflict detection and make it configurable
For reasons lost to the mists of time, this was removed a few years back
in c2fe45f.  We are restoring that code, and making it configurable,
though enabled by default.

This property is a global property; we can't imagine a scenario where a
user would want more fine-grained control over IP conflict detection.

IP conflict detection works by querying the vCenter for VMs on the same
networks with the same IP address.  It does not ping the IP address.

[#186569582]

Co-authored-by: Chris Selzo <chris.selzo@broadcom.com>