IP address conflicts lists ALL conflicting VMs
Previously we only listed the first VM that had a conflicting IP address
when deploying; with this change we now list _all_ the VMs with
conflicting IP addresses.

We also make a slight style change: `File.basename(name)` →
`name.split('/')` because the string we were modifying was not a path,
so using the `File` library seemed misleading.

A more sophisticated question might be, "Why are you splitting on `/`?
Doesn't vSphere allow slashes in port group names? You might be
splitting on the wrong '/'", to which we answer:

Yes, vSphere allows '/', but it URL-encodes them, so when using a CLI
such as `govc` or the MOB browser, you see the encoding:

e.g.

```
govc ls /dc/network
  /dc/network/%2fa%2fb%2fc%2f # NOT "/dc/network//a/b/c/"
```

[#186569582](https://www.pivotaltracker.com/story/show/186569582)

Signed-off-by: Brian Cunnie <brian.cunnie@broadcom.com>