Remove Management API-only NSX-T CI lifecycle jobs
We need to rewrite our NSX-T lifecycle tests to exclusively use the
NSX Policy API because the upcoming NSX 5.0 does not support the
Management API, and 31/37 of our lifecycle tests depend on the
Management API for setup (we don't need to change our code, but we need
to change our tests).

NSX-T 2.5 and earlier _only_ support the Management API, hence we must
deprecate testing against them.

Also, we upgrade the environment we're running BATS against from
vSphere 6.5+NSX-T 2.5 to vSphere 7.0+NSX-T 3.0. I would've upgraded it
further, but I don't think the 8.0 environments are complete enough to
run BATS.

These are the vCenters and NSX-Ts that we are no longer testing against;
they are all past technical guidance:

- NSX-T 2.5 EOGS (end of general support) was 2021-09-19, EOTS (end of technical guidance) was 2022-09-19.
- vCenter 6.5 EOGS was 2022-10-15, EOTS, 2023-11-15
- vCenter 6.7 EOGS was 2022-10-15, EOTS, 2023-11-15