Moves to GitHub & vcpi-pool from master to main
This commit updates the `vsphere-cpi` pipeline to look to GitHub for its
Git resources, rather than in GitLab. The self-hosted GitLab is going
away, so we're moving all this out to GitHub.

We also took the opportunity to rename the master branch in vcpi-pool,
so there's that change in here, too.

Signed-off-by: Brian Cunnie <brian.cunnie@broadcom.com>