Adds plugin system, hook points, supporting libs
This commit adds support for "registering" Ruby plugin code to be called
at the beginning or end of the various CPI v2 method-handling functions.
While any BOSH CPI could use plugins, only the vSphere CPI has been
modified to handle plugins and call plugin hook sites.

This commit also adds what should be a fully-functional toy plugin in
the 'src/vpshere_cpi/src/cpi_plugins/example' directory. This plugin
serves as both a demonstration to vSphere CPI plugin authors, and as a
mechanism for the CPI Integration tests to use to ensure that the plugin
load and call machinery is functioning correctly.

Because we currently have only one user of this plugin mechanism, we are
very, very open to changes to the plugin code, the data passed in to the
various hook sites, the location of said sites, and so on. What we have
is good enough for our uses, but it's hard to design an actually-good
interface when one only has a single user.

For more information on vSphere CPI plugins, their use, and their
configuration, examine
'src/vsphere_cpi/src/cpi_plugins/example/README.md', added by this
commit.

Signed-off-by: Brian Cunnie <brian.cunnie@broadcom.com>
Signed-off-by: Rajath Agasthya <rajath.agasthya@broadcom.com>
Signed-off-by: Nitin Ravindran <nitin.ravindran@broadcom.com>
Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>