Generate multiple IPs when deploying multiple VMs
The "creates more than 5 VMs" NSXT spec was re-using the same random IP
address for all 6 VMs it was deploying. With the re-introduction of IP
conflict detection, this will frequently fail. It seems that under
certain conditions, the vCenter API will not immediately return a VM by
its IP address after it is deployed, allowing the test to sometimes
succeed.

This commit attempts to generate a random set of IP addresses for each
VM. Collisions may still be possible, but should be less frequent.