Modified global cpi config to accept vmx_options like vm_type
Technically this is breaking but in practice this feature has literally
never worked so is it really breaking? Does a tree make a sound if theres
no one around? No.

This should simplify the code paths so disk.enableUUID works the same no matter
if its specified locally or globally.

Co-authored-by: Julian Hjortshoj <julian.hjortshoj@broadcom.com>
Co-authored-by: Nishad Mathur <nishad.mathur@broadcom.com>