fix migration mode code for vm tags
Set tags on ports in manager mode first, so that manager mode changes dont step on the policy mode tags when we set them.
Note: this change fixes a race condition in the tests, we think, but the changes are based on observation and speculation about what NSXT might be doing in 3.2, so we aren't positive that we haven't just changed the race condition and not actually fixed it. Since migration mode will eventually go away, we hope this will be good enough.

Co-authored-by: Nishad Mathur <nishad.mathur@broadcom.com>