fix vsan datastore renaming issue
Previously when detaching a disk in a renamed vsan datastore, we erroneously located the underlying storage file for the disk, not the friendly symlinked location for the same file, and then attempted to move the latter to the location of the former, which failed since there was already a file in that location. This commit adds a check to skip the move operation when the destination file exists and the datastore is of type vsan.
Also, various minor rubocop formatting fixes.

Co-authored-by: Nishad Mathur <nishad.mathur@broadcom.com>