Reaarange host group tests for more robustness
Previously the tests for memory capacity used a host group and a
resource pool with very similar capacity, which means that in VC8 the
tests started failing because nsx-edge is using up some of the memory in
the second cluster, changing the outcome of the VM placement. Switching
the test around to use a resource pool that is considerably bigger
should make the tests more predictable.

Co-authored-by: Colin Shield <colin.shield@broadcom.com>